Refactror
